requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.6.0
marshmallow==3.21.0
marshmallow-sqlalchemy==0.29.0
Werkzeug==3.0.1
gunicorn==21.2.0
pytest==7.4.3
pytest-flask==1.3.0

# MySQL dependencies
PyMySQL==1.1.0
cryptography==41.0.7

# Additional utilities
bcrypt==4.1.2
email-validator==2.1.0
python-dateutil==2.8.2